Each policy type has a separated class
