+++ projects/com.nvidia.grcuda.test/src/com/nvidia/grcuda/test/runtime/executioncontext/GrCUDAStreamPolicyMockTest.java
@@ -7,6 +7,8 @@ import com.nvidia.grcuda.runtime.Device;
 import com.nvidia.grcuda.runtime.computation.dependency.DependencyPolicyEnum;
 import com.nvidia.grcuda.runtime.stream.policy.DeviceSelectionPolicyEnum;
 import com.nvidia.grcuda.runtime.stream.policy.GrCUDAStreamPolicy;
+import com.nvidia.grcuda.runtime.stream.policy.RoundRobinDeviceSelectionPolicy;
+import com.nvidia.grcuda.runtime.stream.policy.TransferTimeDeviceSelectionPolicy;
 import com.nvidia.grcuda.runtime.stream.policy.RetrieveNewStreamPolicyEnum;
 import com.nvidia.grcuda.runtime.stream.policy.RetrieveParentStreamPolicyEnum;
 import com.nvidia.grcuda.test.util.mock.AsyncGrCUDAExecutionContextMock;
@@ -41,14 +43,14 @@ public class GrCUDAStreamPolicyMockTest {
         );
     }
 
-    private GrCUDAStreamPolicy.RoundRobinDeviceSelectionPolicy getRoundRobinPolicy(int numGPUs) {
+    private RoundRobinDeviceSelectionPolicy getRoundRobinPolicy(int numGPUs) {
         GrCUDADevicesManagerMock devicesManager = new GrCUDADevicesManagerMock(new DeviceListMock(numGPUs), numGPUs);
-        return new GrCUDAStreamPolicy.RoundRobinDeviceSelectionPolicy(devicesManager);
+        return new RoundRobinDeviceSelectionPolicy(devicesManager);
     }
 
     @Test
     public void roundRobinTest() {
-        GrCUDAStreamPolicy.RoundRobinDeviceSelectionPolicy policy = getRoundRobinPolicy(4);
+        RoundRobinDeviceSelectionPolicy policy = getRoundRobinPolicy(4);
         Device d = policy.retrieve(null);
         assertEquals(0, d.getDeviceId());
         assertEquals(1, policy.getInternalState());
@@ -126,7 +128,7 @@ public class GrCUDAStreamPolicyMockTest {
                 {45, 60, 20},
                 {10, 20, 0}
         };
-        double[][] b = ((GrCUDAStreamPolicy.TransferTimeDeviceSelectionPolicy) streamPolicy.getDeviceSelectionPolicy()).getLinkBandwidth();
+        double[][] b = ((TransferTimeDeviceSelectionPolicy) streamPolicy.getDeviceSelectionPolicy()).getLinkBandwidth();
         for (int i = 0; i < b.length; i++) {
             for (int j = 0; j < b[i].length; j++) {
                 assertEquals(bGold[i][j], b[i][j], 1e-6);
@@ -137,7 +139,7 @@ public class GrCUDAStreamPolicyMockTest {
     @Test
     public void bandwidthComputationMinMaxTest() {
         AsyncGrCUDAExecutionContextMock context = createContext(2, DeviceSelectionPolicyEnum.MINMAX_TRANSFER_TIME);
-        GrCUDAStreamPolicy.TransferTimeDeviceSelectionPolicy deviceSelectionPolicy = (GrCUDAStreamPolicy.TransferTimeDeviceSelectionPolicy) ((GrCUDAStreamPolicyMock) context.getStreamManager().getStreamPolicy()).getDeviceSelectionPolicy();
+        TransferTimeDeviceSelectionPolicy deviceSelectionPolicy = (TransferTimeDeviceSelectionPolicy) ((GrCUDAStreamPolicyMock) context.getStreamManager().getStreamPolicy()).getDeviceSelectionPolicy();
         // If data is updated on the target device, we have infinite bandwidth (regardless of what's on the matrix diagonal);
         double b = deviceSelectionPolicy.computeBandwidth(0, new HashSet<>(Arrays.asList(0, 1, CPUDevice.CPU_DEVICE_ID)));
         assertEquals(Double.POSITIVE_INFINITY, b, 1e-6);
@@ -149,7 +151,7 @@ public class GrCUDAStreamPolicyMockTest {
     @Test
     public void bandwidthComputationMinMinTest() {
         AsyncGrCUDAExecutionContextMock context = createContext(2, DeviceSelectionPolicyEnum.MINMIN_TRANSFER_TIME);
-        GrCUDAStreamPolicy.TransferTimeDeviceSelectionPolicy deviceSelectionPolicy = (GrCUDAStreamPolicy.TransferTimeDeviceSelectionPolicy) ((GrCUDAStreamPolicyMock) context.getStreamManager().getStreamPolicy()).getDeviceSelectionPolicy();
+        TransferTimeDeviceSelectionPolicy deviceSelectionPolicy = (TransferTimeDeviceSelectionPolicy) ((GrCUDAStreamPolicyMock) context.getStreamManager().getStreamPolicy()).getDeviceSelectionPolicy();
         // If data is updated on the target device, we have infinite bandwidth (regardless of what's on the matrix diagonal);
         double b = deviceSelectionPolicy.computeBandwidth(0, new HashSet<>(Arrays.asList(0, 1, CPUDevice.CPU_DEVICE_ID)));
         assertEquals(Double.POSITIVE_INFINITY, b, 1e-6);
@@ -161,7 +163,7 @@ public class GrCUDAStreamPolicyMockTest {
     @Test(expected = IllegalStateException.class)
     public void bandwidthComputationWithNoUpdatedLocationTest() {
         AsyncGrCUDAExecutionContextMock context = createContext(2, DeviceSelectionPolicyEnum.MINMAX_TRANSFER_TIME);
-        GrCUDAStreamPolicy.TransferTimeDeviceSelectionPolicy deviceSelectionPolicy = (GrCUDAStreamPolicy.TransferTimeDeviceSelectionPolicy) ((GrCUDAStreamPolicyMock) context.getStreamManager().getStreamPolicy()).getDeviceSelectionPolicy();
+        TransferTimeDeviceSelectionPolicy deviceSelectionPolicy = (TransferTimeDeviceSelectionPolicy) ((GrCUDAStreamPolicyMock) context.getStreamManager().getStreamPolicy()).getDeviceSelectionPolicy();
         // If the data is not available on any device, give an error;
         double b = deviceSelectionPolicy.computeBandwidth(0, new HashSet<>());
     }
@@ -169,7 +171,7 @@ public class GrCUDAStreamPolicyMockTest {
     @Test(expected = ArrayIndexOutOfBoundsException.class)
     public void bandwidthComputationOutOfBoundsLocationTest() {
         AsyncGrCUDAExecutionContextMock context = createContext(2, DeviceSelectionPolicyEnum.MINMAX_TRANSFER_TIME);
-        GrCUDAStreamPolicy.TransferTimeDeviceSelectionPolicy deviceSelectionPolicy = (GrCUDAStreamPolicy.TransferTimeDeviceSelectionPolicy) ((GrCUDAStreamPolicyMock) context.getStreamManager().getStreamPolicy()).getDeviceSelectionPolicy();
+        TransferTimeDeviceSelectionPolicy deviceSelectionPolicy = (TransferTimeDeviceSelectionPolicy) ((GrCUDAStreamPolicyMock) context.getStreamManager().getStreamPolicy()).getDeviceSelectionPolicy();
         // If the data is not available on any device, give an error;
         double b = deviceSelectionPolicy.computeBandwidth(10, new HashSet<>(Collections.singletonList(1)));
     }
